backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
deepface==0.0.93
opencv-python==4.10.0.84
numpy==1.26.4
python-multipart==0.0.12
Pillow==10.4.0
tf-keras==2.17.0
